requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alpaca-py==0.37.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
Backtesting==0.5.0
bokeh==3.6.2
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
deprecation==2.1.0
executing==2.2.0
fonttools==4.55.7
frozenlist==1.5.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
msgpack==1.0.3
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sseclient-py==1.8.0
stack-data==0.6.3
ta-lib==0.6.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
xyzservices==2025.1.0
yarl==1.18.3
